Main Menu State
Added a Main Menu State including placeholder buttons. No functionality yet.

Also finished up the Splash State by adding a close window event for if the user closes the window. And made it so that the Splash State changes to the Main Menu state after a defined amount of time.