World Map
Added the World Map state, including an init() implementation which sets up the vertex array for all the quads (tiles) of the map and defines a tile matrix. Also colors them in with placeholder colors so that they are visible.

Also added related constants to defines, and implemented the state transition from Main Menu to World Map.