Fix input manager and camera to world map
Made the implementation of the isSpriteClicked()-function neater and more consistent by making it check the mouse coordinates vs the global bounds of the sprite directly instead of  doing a roundabout check like it did before, which also caused bugs if the origin point of the sprite had not been set to the top left corner.

Implemeneted a "camera" to the world map using the sf::View class and added basic navigation controls to it, including boundaries which the camera can not bypass.