Tile clicking system
Messed around with a different map for testing purposes, nothing to worry about!

But I also added a system for registering when a specific tile is clicked, regardless of the position or zoom of the camera. I just added some testing stuff for now (writing out tile data in the console for the tile that was clicked) but I'll expand upon it next!