Tile Data Box State
Added a new state class for the Tile Data Box, which is set to appear when a tile is clicked on the world map. The tile data box contains terrain data for that particular tile. Currently the tile data box state completely overrides the world map state so that it isn't shown or anything, but ideally I'd like to rework the game engine to allow the box state to appear alongside the world map without overriding it completely.

Also added various resources related to the tile data box, such as textures and fonts. Also defines etc.

Very incomplete though, for instance there's no way to go back to the world map currently.