Reversed the order of substate updates.
Changed it so that the substates are updated from top to bottom rather than from bottom to top, since that's probably what you'd want. This included changing the iterator to a reverse_iterator, which was pretty straight-forward except for the erase() function which made everything fiendishly complicated all of a sudden, but in the end I learned a bunch of new shit, such as what decltype is and how it can be used!