Added separate view for data boxe substate.
Added a separate view for the Tile Data Box substate which works independently from the world camera view. That way the world camera can move and zoom while the data boxes stay at the same position relative to the window, and also ensures that the position of for instance the close button remains fixed at the location it visually appears to be in.