Selected tile data boxes now have an alternate texture.
Added an alternate texture for the tile data box to be used for whichever box is currently selected. Also added functionality to switch to said texture whenever a data box is on top of the substate list, and otherwise switch to the default texture when a given data box is not on top of the substate list.

I kinda wonder how much of this is superfluous though, since I'm not even sure if it would be desirable in the end to be able to have multiple tile data boxes open at the same time. Oh well.