Added a main GameState and restructured/renamed a bunch of code to fit this new design.
Created a main GameState which contains its own StateMachine to handle states that are specific to the game itself (worldmap, regional maps, etc). The engine now adds the GameState to its StateMachine stack and then the GameState adds the World Map and all updates, input handling and drawing of the maps is handled through the GameState.

Also renamed a few things, such as the GameData struct which is now called "EngineData", since there's now a new GameData struct which is used for the internal game database.

I'm not sure if this restructuring is actually a good idea, it seemed to make sense in some way but I'm not 100% sure, we'll see I guess.