Implemented a HUD class (barebones as of now).
Implemented a HUD-class and added it as a member to GameState. Also changed how the game handles drawing GameState and other states within it, namely that the GameState draw-function handles clearing and displaying the window by itself while all states below just handles drawing themselves.

The HUD gets drawn after all other states so it always ends up on top.

I'm really not sure if this implementation of a HUD is good, nor am I sure if the structure of the GameState class and the idea of it containing all other states is a good idea. We'll see.