Renamed Engine Data stuff to more appropriate names.
Renamed a bunch of stuff relating to the engine, like GameDataRef -> EngineDataRef, _data -> _engine, and so on. The reason is that I'm probably going to add a similar database for game-specific stuff which it would probably make sense to call "Game Data" so the engine stuff would more appropriately be called "Engine Data".

Also added the file for the RegionMap class but haven't really done anything with it yet.