Implementation of the RegionMap class (inflexible as of now).
Implemented the RegionMap class as essentially a copy of the WorldMap class but with it's own Define constants for most things. Also updated the TileDataBox class so it has a Enter button to change from the WorldMap to the RegionMap.

Right now it's all very inflexible though, as there's only one predefined region map of one specific region of the world. I'd probably need some kind of automated map generator to make it more flexible though.