Added some game database stuff + restructured filters
Added a few structs and classes meant to be used as the global database containing all the world data pertaining to a individual game. This includes a template struct called Property which represents, well, a property of another database struct/class, and has different values for its actual value and for the range of possible values displayed to the player based on their current knowledge of the thing.

I also restructured some filters in visual studio which maybe doesn't even affect anything in the actual file strucutre.