Setting Box Tab labels
Added constructors and functions for setting the labels of tabs in a box. They can be set by passing a initializer list with strings to the Box, which will then dynamically resize the Tabs array to the number of tabs required to fit the list. Then each label is passed to each tab in turn.