BoxTabs from vector to map
Changed the BoxTab container in the Box class from a std::vector to a std::map using the displayed tab labels as map keys, since it seems more intuitive that way. Changed all relevant functions in Box to accomodate this change, including keeping a separate vector for just the labels so that the tabs can be positioned in the desired order (instead of the built-in ordering of std::map).