setPosition for BoxTab and fixed positioning flaws
Added functions for setting the tab position and content area position for BoxTab, in addition to the movePosition function that was already there, since I realized a function for setting the initial positions are needed.

Also made the content area visible for now so that I can confirm that it has the proper size and position. It turns out it didn't before, so I also fixed that.