_contentArea moved from Box to BoxTab, type changed from RectangleShape to FloatRect
Made _contentArea a member of Box instead of BoxTab since in the end it just made more sense (each tab should have the same content area anyway, and that area would be based on the size of the box as a whole among other things). Also changed it from a RectangleShape to a FloatRect since it's really only meant to define an area, not display a visual shape.