Tab content scrolling functionality.
Added scrolling functionality for the tabContentView by dragging the scroll handle on the Box. This includes member variables and functions within BoxTab for keeping track of the top and bottom relative Y-coordinates of the tab contents, so that the camera scrolls from the top to the bottom using the relative position of the scroll handle to the scroll bar.

This has some bugginess going on at the moment, so it could do with some cleanup.