Database structs use database indexes instead of pointers to other database items
Altered the database structs for minerals, stones and bedrocks so that they keep track of the database index or their respective mineralType/stoneType rather than a pointer to it. I think this makes more sense as it allows for using the database index as an ID-number for each item. An alternative would've been to have an ID-number as a member variable for each struct, but I dunno I think this works better maybe, we'll see.

Also fixed the calculateStoneProperties function so that it accurately produces the average mineral properties.