Made the subState list into a member of the State base class.
Made it so that the subState list is a member of the State base class and also implemented a getter function for that list. This should allow for instance subStates of WorldMap to access the subState list of WorldMap and add substates to it. This could be relevant to for instance allow a TileDataBox to open a DatabaseBox if an item (for instance a Mineral) in the TileDataBox is clicked.