Substate list is now shared pointers instead of unique pointers, and makeLink() now works.
- Changed the subStateRef from a unique_ptr to a shared_ptr, which should be functionally identical in most situations but it does allow for storing a substate somewhere else and then passing it over to the substate list to be displayed/handled, while still retaining the substate in the other place for future reference.
Right now I'm using this to have the DatabaseBox class store links to one MineralBox for each mineral type listed in the DatabaseBox, so that when a mineral type is clicked in the list the associated stored MineralBox is passed over to the subState list in WorldMap to be displayed, but it's kept stored in DatabaseBox so that it can be clicked again. If that makes sense!

- Implemented a working version of the makeLink() function in BoxTab thanks to the above changes and made DatabaseBox use it for links to the MineralBoxes associated with the mineral types listed in the DatabaseBox, as explained above.